Revert "ld_align"
This reverts commit c474e75b4081a7cbaf2188b66b704fd47fec03f0.

+++ lasercom/polarizer/IPolarizer.cs
@@ -52,7 +52,7 @@
         /// <summary>
         /// Current beta setting of polarizers.
         /// </summary>
-        float PolarizerBeta { get; set; }
+        int PolarizerBeta { get; set; }
 
         /// <summary>
         /// Min beta setting of polarizers.
